+++ src/common/classes/command-with-redirect-to-filepath.ts
@@ -1,6 +1,5 @@
+import { execCommand } from '../exec-command';
 import { JSONSchemaType } from 'ajv';
-import { promisify }      from 'util';
-import { exec }           from 'child_process';
 
 import { FilePathGlobalOptionSchema, SchemaIds } from '../constants/schemas';
 import { CommandError }                          from '../errors/command';
@@ -8,7 +7,6 @@ import { GlobalOptionsWithRequiredFilePath }     from '../interfaces/common';
 import Command                                   from './command';
 import { validate }                              from '../misc';
 
-const promisifiedExec = promisify(exec);
 
 export default class CommandWithRedirectToFilepath<T_CommandOptions extends { [index: string]: any; }> extends Command<T_CommandOptions> {
   constructor(protected schemaId: SchemaIds,
@@ -39,7 +37,7 @@ export default class CommandWithRedirectToFilepath<T_CommandOptions extends { [i
       return this;
     }
 
-    const { stderr } = await promisifiedExec(this._cmdToExec);
+    const { stderr } = await execCommand(this._cmdToExec);
 
     if (!stderr) {
       return this;